Rename cc to CaesarCipher
